Split view tests into separate files.
